Add golangci-lint and gofumpt
The linter helps catching common errors, security problems and style.
For now the `annoying` linters are disabled but you should think about
enabling those who work for you.

Gofumpt can be replaced with gofmt ofc, but it is a bit stricter, which
is really nice.

The Makefile is just a command runner, because that is usually installed
anyway.