Fix net/http.Server security vulnerability
Not setting a ReadHeaderTimeout could lead to timeout attacks
and therefore make a server unresponsive.

This was detected by golangci-lint.
I choose a value derived from Nginx's and Apache's default configuration.
Although, it now is a magic number,
I don't think it needs to be configurable,
since the server will be overwritten by a custom server anyway.